refactor replaceXAudioDllsFromRedistributable to XServerScreenUtils
add other gamesource detection
