+++ app/src/main/java/app/gamenative/utils/SteamUtils.kt
@@ -1494,89 +1494,5 @@ object SteamUtils {
             Timber.e(e, "Failed to generate achievements for $appId")
         }
     }
-
-    /**
-     * Replace DLLs from DirectX Redistributable
-     */
-    fun replaceXAudioDllsFromRedistributable(context: Context, appId: String) {
-        val imageFs = ImageFs.find(context)
-        val rootDir = imageFs.rootDir
-        val windowsDir = File(rootDir, ImageFs.WINEPREFIX + "/drive_c/windows")
-        val steamAppId = ContainerUtils.extractGameIdFromContainerId(appId)
-        val gameDirPath = SteamService.getAppDirPath(steamAppId)
-        val directXDir = File(gameDirPath, "_CommonRedist/DirectX")
-
-        // Do this once at app startup or before the loop
-        SevenZip.initSevenZipFromPlatformJAR()
-
-        fun extractDllsFromCab(cabFile: File, targetDir: File) {
-            val raf = RandomAccessFile(cabFile, "r")
-            val inStream = RandomAccessFileInStream(raf)
-            val archive: IInArchive = SevenZip.openInArchive(null, inStream)
-
-            try {
-                val numberOfItems = archive.numberOfItems
-                for (i in 0 until numberOfItems) {
-                    val name = archive.getProperty(i, net.sf.sevenzipjbinding.PropID.PATH) as String
-
-                    if (name.endsWith(".dll", ignoreCase = true)) {
-                        val outFile = File(targetDir, name.lowercase())
-                        if (outFile.exists()) {
-                            outFile.delete()
-                        }
-
-                        archive.extract(
-                            intArrayOf(i), false,
-                            object : IArchiveExtractCallback {
-                                override fun getStream(index: Int, extractAskMode: ExtractAskMode): ISequentialOutStream? {
-                                    if (extractAskMode != ExtractAskMode.EXTRACT) return null
-
-                                    return ISequentialOutStream { data ->
-                                        FileOutputStream(outFile, true).use { it.write(data) }
-                                        data.size // Return the number of bytes written
-                                    }
-                                }
-
-                                override fun prepareOperation(extractAskMode: ExtractAskMode) {}
-                                override fun setOperationResult(extractOperationResult: ExtractOperationResult) {}
-                                override fun setCompleted(completeValue: Long) {}
-                                override fun setTotal(total: Long) {}
-                            }
-                        )
-
-                        Timber.tag("replaceXAudioDllsFromRedistributable").d("Extracted: ${outFile.name}")
-                    }
-                }
-            } finally {
-                archive.close()
-                inStream.close()
-                raf.close()
-            }
-        }
-
-        if (directXDir.exists()) {
-            directXDir.walkTopDown()
-                .filter { file ->
-                    val name = file.name.lowercase()
-                    val isAudioType =
-                        name.contains("xaudio") ||
-                        name.contains("xact") ||
-                        name.contains("x3daudio")
-
-                    isAudioType && file.extension.equals("cab", ignoreCase = true)
-                }
-                .forEach { cabFile ->
-                    Timber.tag("replaceXAudioDllsFromRedistributable").d("Processing cabinet: ${cabFile.name}")
-
-                    if (cabFile.name.lowercase().contains("x86")) {
-                        val targetDir = File(windowsDir, "syswow64")
-                        extractDllsFromCab(cabFile, targetDir)
-                    } else if (cabFile.name.lowercase().contains("x64")) {
-                        val targetDir = File(windowsDir, "system32")
-                        extractDllsFromCab(cabFile, targetDir)
-                    }
-                }
-        }
-    }
 }
